degrade vector and matrix
