+++ src/for_2D_build/common/data_type.h
@@ -31,7 +31,6 @@
 
 #include "base_data_type.h"
 #include "scalar_functions.h"
-#include "vector_functions.h"
 
 namespace SPH
 {
