+++ src/for_3D_build/common/data_type.h
@@ -30,7 +30,6 @@
 
 #include "base_data_type.h"
 #include "scalar_functions.h"
-#include "vector_functions.h"
 
 namespace SPH
 {
