+++ src/for_3D_build/common/vector_functions_supplementary.cpp
@@ -1,0 +1,16 @@
+#include "vector_functions.h"
+
+namespace SPH
+{
+//=================================================================================================//
+Vec3d degradeToVecd(const Vec3d &input)
+{
+    return input;
+}
+//=================================================================================================//
+Mat3d degradeToMatd(const Mat3d &input)
+{
+    return input;
+}
+//=================================================================================================//
+} // namespace SPH
