+++ src/shared/common/base_data_package.h
@@ -32,6 +32,7 @@
 #include "data_type.h"
 #include "large_data_containers.h"
 #include "ownership.h"
+#include "vector_functions.h"
 
 #define TBB_PARALLEL true
