+++ src/shared/common/vector_functions.cpp
@@ -40,19 +40,6 @@ Mat3d upgradeToMat3d(const Mat3d &input)
     return input;
 }
 //=================================================================================================//
-void degradeToVecd(const Vec3d &input, Vec2d &output)
-{
-    output[0] = input[0];
-    output[1] = input[1];
-}
-//=================================================================================================//
-void degradeToMatd(const Mat3d &input, Mat2d &output)
-{
-    for (int i = 0; i != 2; i++)
-        for (int j = 0; j != 2; j++)
-            output(i, j) = input(i, j);
-}
-//=================================================================================================//
 Mat2d getInverse(const Mat2d &A)
 {
     Mat2d minv = Mat2d::Zero();
@@ -205,7 +192,7 @@ Mat3d getTransformationMatrix(const Vec3d &direction_of_z, const Vec3d &directio
     transformation_matrix.row(0) = direction_of_x.transpose();
     transformation_matrix.row(1) = direction_of_y.transpose();
     transformation_matrix.row(2) = direction_of_z.transpose();
-     
+
     return transformation_matrix;
 }
